requirements.txt
fastapi==0.115.12
langchain-anthropic==0.3.12
langchain-community==0.3.23
langchain-core==0.3.56
langchain-mcp-adapters==0.0.9
langchain-openai==0.3.15
langchain-pinecone==0.2.5
langgraph==0.4.1
langgraph-supervisor==0.0.20
langmem==0.0.25
markdown==3.8
nats-py[nkeys]==2.10.0
numpy==2.2.5
openpyxl==3.1.5
pinecone[asyncio,grpc]==6.0.2
psycopg==3.2.7
psycopg2-binary==2.9.10
pydantic==2.11.4
pymongo==4.12.1
pytest==8.3.5
python-ulid==3.0.0
trustcall==0.0.39
unstructured==0.17.2
uvicorn==0.34.2
